+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* remove Dependencies tab, Manage Dependencies link, etc, per @azaozz
 
 #### 1.14.3 / 2023-70-30
 * add null coalesce